+++ stac_fastapi/extensions/stac_fastapi/extensions/core/filter/request.py
@@ -1,8 +1,8 @@
 """Filter extension request models."""
 
-from dataclasses import dataclass
 from typing import Any, Dict, Literal, Optional
 
+import attr
 from fastapi import Query
 from pydantic import BaseModel, Field
 from typing_extensions import Annotated
@@ -12,13 +12,17 @@ from stac_fastapi.types.search import APIRequest
 FilterLang = Literal["cql-json", "cql2-json", "cql2-text"]
 
 
-@dataclass
+@attr.s
 class FilterExtensionGetRequest(APIRequest):
     """Filter extension GET request model."""
 
-    filter: Annotated[Optional[str], Query()] = None
-    filter_crs: Annotated[Optional[str], Query(alias="filter-crs")] = None
-    filter_lang: Annotated[Optional[FilterLang], Query(alias="filter-lang")] = "cql2-text"
+    filter: Annotated[Optional[str], Query()] = attr.ib(default=None)
+    filter_crs: Annotated[Optional[str], Query(alias="filter-crs")] = attr.ib(
+        default=None
+    )
+    filter_lang: Annotated[Optional[FilterLang], Query(alias="filter-lang")] = attr.ib(
+        default="cql2-text"
+    )
 
 
 class FilterExtensionPostRequest(BaseModel):