+++ stac_fastapi/extensions/stac_fastapi/extensions/core/pagination/request.py
@@ -1,8 +1,8 @@
 """Pagination extension request models."""
 
-from dataclasses import dataclass
 from typing import Optional
 
+import attr
 from fastapi import Query
 from pydantic import BaseModel
 from typing_extensions import Annotated
@@ -10,11 +10,11 @@ from typing_extensions import Annotated
 from stac_fastapi.types.search import APIRequest
 
 
-@dataclass
+@attr.s
 class GETTokenPagination(APIRequest):
     """Token pagination for GET requests."""
 
-    token: Annotated[Optional[str], Query()] = None
+    token: Annotated[Optional[str], Query()] = attr.ib(default=None)
 
 
 class POSTTokenPagination(BaseModel):
@@ -23,11 +23,11 @@ class POSTTokenPagination(BaseModel):
     token: Optional[str] = None
 
 
-@dataclass
+@attr.s
 class GETPagination(APIRequest):
     """Page based pagination for GET requests."""
 
-    page: Annotated[Optional[str], Query()] = None
+    page: Annotated[Optional[str], Query()] = attr.ib(default=None)
 
 
 class POSTPagination(BaseModel):