+++ stac_fastapi/extensions/stac_fastapi/extensions/core/query/request.py
@@ -1,8 +1,8 @@
 """Request model for the Query extension."""
 
-from dataclasses import dataclass
 from typing import Any, Dict, Optional
 
+import attr
 from fastapi import Query
 from pydantic import BaseModel
 from typing_extensions import Annotated
@@ -10,11 +10,11 @@ from typing_extensions import Annotated
 from stac_fastapi.types.search import APIRequest
 
 
-@dataclass
+@attr.s
 class QueryExtensionGetRequest(APIRequest):
     """Query Extension GET request model."""
 
-    query: Annotated[Optional[str], Query()] = None
+    query: Annotated[Optional[str], Query()] = attr.ib(default=None)
 
 
 class QueryExtensionPostRequest(BaseModel):