+++ stac_fastapi/extensions/stac_fastapi/extensions/core/sort/request.py
@@ -1,9 +1,8 @@
-# encoding: utf-8
 """Request model for the Sort Extension."""
 
-from dataclasses import dataclass
 from typing import List, Optional
 
+import attr
 from fastapi import Query
 from pydantic import BaseModel
 from stac_pydantic.api.extensions.sort import SortExtension as PostSortModel
@@ -12,16 +11,11 @@ from typing_extensions import Annotated
 from stac_fastapi.types.search import APIRequest, str2list
 
 
-@dataclass
+@attr.s
 class SortExtensionGetRequest(APIRequest):
     """Sortby Parameter for GET requests."""
 
-    sortby: Annotated[Optional[str], Query()] = None
-
-    def __post_init__(self):
-        """convert attributes."""
-        if self.sortby:
-            self.sortby = str2list(self.sortby)  # type: ignore
+    sortby: Annotated[Optional[str], Query()] = attr.ib(default=None, converter=str2list)
 
 
 class SortExtensionPostRequest(BaseModel):